Refactor pathTokenAuthorizer middleware into file module
